meeting-summary-bot/requirements.txt
anthropic>=0.25.0
openai>=1.30.0
python-dotenv>=1.0.0
